Cache language in session
