+++ app/misc.py
@@ -25,7 +25,7 @@ from slugify import slugify as s_slugify
 
 from .config import config
 from flask_login import AnonymousUserMixin, current_user
-from flask_babel import _
+from flask_babel import Babel, _
 from flask_talisman import Talisman
 from .caching import cache
 from .socketio import socketio
@@ -77,6 +77,7 @@ engine = _engine
 
 mail = Mail()
 
+babel = Babel()
 
 talisman = Talisman()
 
@@ -832,6 +833,11 @@ def load_user(user_id):
                      on=(Notification.target == User.uid) & Notification.read.is_null(True)).switch(User)
     user = user.group_by(User.uid).where(User.uid == user_id).dicts().get()
 
+    # This is the only user attribute needed by the error templates, so stash
+    # it in the session so that future errors in this session won't have to
+    # load the user to show them the correct language.
+    session['language'] = user['language']
+
     if request.path == '/socket.io/':
         return SiteUser(user, [], [])
     else:
@@ -845,6 +851,29 @@ def load_user(user_id):
             return SiteUser(user, subs, prefs)
         except User.DoesNotExist:
             return None
+
+
+def user_is_loaded():
+    return has_request_context() and hasattr(_request_ctx_stack.top, 'user')
+
+
+def ensure_locale_loaded():
+    if 'language' not in session or not session['language']:
+        session['language'] = get_locale_fallback()
+
+
+@babel.localeselector
+def get_locale():
+    language = session.get('language', None)
+    if language:
+        return language
+    if current_user.language:
+        return current_user.language
+    return get_locale_fallback()
+
+
+def get_locale_fallback():
+    return request.accept_languages.best_match(config.app.languages, config.app.fallback_language)
 
 
 def get_notification_count(uid):
@@ -1856,7 +1885,7 @@ def add_context_to_log_records(config):
                         record.__dict__[k] = unavailable
                 elif var == 'current_user':
                     # Peek at the current user but don't load it if not loaded.
-                    if has_request_context() and hasattr(_request_ctx_stack.top, 'user'):
+                    if user_is_loaded():
                         record.__dict__[k] = getattr(current_user, attr, unavailable)
                     else:
                         record.__dict__[k] = unavailable
